refactor: unify symbol normalization
