+++ types/converters/symbols.ts
@@ -139,67 +139,6 @@ export function toDbSymbol(symbol: SymbolInfo): DbSymbolRecord {
     min_qty: symbol.minQty,
     max_qty: symbol.maxQty,
     status: symbol.status || 'TRADING',
-    is_active: symbol.isActive !== false
+  is_active: symbol.isActive !== false
   };
 }
-
-/**
- * シンボル文字列をベースアセットとクォートアセットに分解
- * @param symbol シンボル文字列 (例: "BTCUSDT")
- * @returns [baseAsset, quoteCoin]のタプル
- */
-export function splitSymbol(symbol: string): [string, string] {
-  // 一般的なクォート通貨リスト（長いものから優先的にマッチ）
-  const commonQuotes = ['USDT', 'BUSD', 'USDC', 'USD', 'BTC', 'ETH', 'BNB'];
-  
-  // シンボルを大文字に変換して処理
-  const upperSymbol = symbol.toUpperCase();
-  
-  // 既知のクォート通貨でマッチングを試みる
-  for (const quote of commonQuotes) {
-    if (upperSymbol.endsWith(quote)) {
-      const base = upperSymbol.slice(0, upperSymbol.length - quote.length);
-      if (base.length > 0) {
-        return [base, quote];
-      }
-    }
-  }
-  
-  // マッチしない場合は経験則ベースで分割を試みる
-  // BTC/USDTのようなスラッシュ形式
-  if (upperSymbol.includes('/')) {
-    const [base, quote] = upperSymbol.split('/');
-    return [base, quote];
-  }
-  
-  // 経験則: 通常、クォート通貨は3-4文字
-  const base = upperSymbol.slice(0, -4);
-  const quote = upperSymbol.slice(-4);
-  
-  // ベースが存在しない場合は別の分割を試す
-  if (base.length === 0) {
-    return [upperSymbol.slice(0, -3), upperSymbol.slice(-3)];
-  }
-  
-  return [base, quote];
-}
-
-/**
- * シンボル文字列の標準化（BTCUSDTやBTC/USDTなど異なる形式を標準形式に変換）
- * @param symbol 変換対象のシンボル文字列
- * @param separator ベースとクォートの区切り文字（デフォルトは空文字）
- * @returns 標準化されたシンボル文字列
- */
-export function normalizeSymbol(symbol: string, separator: string = ''): string {
-  if (!symbol) return '';
-  
-  // スラッシュ形式を処理
-  if (symbol.includes('/')) {
-    const [base, quote] = symbol.split('/');
-    return `${base.toUpperCase()}${separator}${quote.toUpperCase()}`;
-  }
-  
-  // 標準形式を文字列分割で推定
-  const [base, quote] = splitSymbol(symbol);
-  return `${base.toUpperCase()}${separator}${quote.toUpperCase()}`;
-}
